Another architectural change.
The constructor-helper has been removed in favor of generalizing the
setters.  This eliminates `bytestructure-fill!'.

The distinction between compound and non-compound descriptors has been
lifted, since the only remaining difference was that non-compound
types were not allowed to have ref-helpers; now they simply are
allowed that as well, although this feature is not used in any
built-in type so far.