Fix union size calculation.
A union's size is the smallest multiple of its alignment that can
contain the biggest element in the union.

Consider your union has a uint32 and a uint16 vector of size 3.  This
union only needs 48 bits of storage, but it must be 64 bits large due
to the 32-bit alignment of the uint32.